backend/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client>=1.8.0
openai-agents
mangum
